+++ lib/lims-laboratory-app/laboratory/plate/create_plate.rb
@@ -1,6 +1,4 @@
 # vi: ts=2:sts=2:et:sw=2  spell:spelllang=en  
-require 'lims-core/actions/action'
-
 require 'lims-laboratory-app/laboratory/plate'
 require 'lims-laboratory-app/laboratory/container/create_container_action'
