+++ spec/models/laboratory/container_like_asset_shared.rb
