+++ spec/models/laboratory/plate/plate_sequel_persistor_spec.rb
@@ -1,7 +1,7 @@
 # Spec requirements
 require 'models/persistence/sequel/spec_helper'
 
-require 'models/laboratory/plate_and_gel_shared'
+require 'models/laboratory/container_like_asset_shared'
 require 'models/persistence/resource_shared'
 require 'models/persistence/sequel/store_shared'
 require 'models/persistence/filter/multi_criteria_sequel_filter_shared'
