+++ spec/models/laboratory/plate/plate_transfer_spec.rb
@@ -1,7 +1,7 @@
 # Spec requirements
 require 'models/actions/spec_helper'
 require 'models/actions/action_examples'
-require 'models/laboratory/plate_and_gel_shared'
+require 'models/laboratory/container_like_asset_shared'
 
 require 'models/persistence/sequel/store_shared'
 #Model requirements
