+++ spec/models/laboratory/tag_wells_spec.rb
@@ -3,7 +3,7 @@ require 'models/actions/spec_helper'
 require 'models/actions/action_examples'
 
 require 'models/persistence/sequel/spec_helper'
-require 'models/laboratory/plate_and_gel_shared'
+require 'models/laboratory/container_like_asset_shared'
 require 'models/persistence/sequel/store_shared'
 
 #Model requirements
